[FIX] queue_job: calling create() with empty values
Change the _job_edit_sentinel check in the override of create() on queue.job so that the exception is not raised when called with an empty value list. This fixes the TestOverrides.test_creates unit test that enforces that create() can be always be called with empty values. (see odoo/odoo#116809)